+++ example/ntt_merge/test_merge_ntt.cu
@@ -59,9 +59,9 @@ int main(int argc, char* argv[])
     std::random_device rd;
     // std::mt19937 gen(rd());
     std::mt19937 gen(0);
-    unsigned long long minNumber = 0;
-    unsigned long long maxNumber = parameters.modulus.value - 1;
-    std::uniform_int_distribution<unsigned long long> dis(minNumber, maxNumber);
+    unsigned long minNumber = 0;
+    unsigned long maxNumber = parameters.modulus.value - 1;
+    std::uniform_int_distribution<unsigned long> dis(minNumber, maxNumber);
 
     // Random data generation for polynomials
     vector<vector<TestDataType>> input1(BATCH);
